Добавил в проект webpack
