+++ src/index.html
@@ -456,6 +456,6 @@
         </div>
     </section>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/tiny-slider/2.9.2/min/tiny-slider.js"></script>
-    <script src="js/script.js"></script>
+    <script src="./js/bundle.js"></script>
 </body>
 </html>
